refactor: [WIP] move icon into local class
