+++ src/react/table-app/option-bar/sort-button.tsx
@@ -2,7 +2,7 @@ import Button from "src/react/shared/button";
 import { SortDir } from "src/data/types";
 import Stack from "src/react/shared/stack";
 import Icon from "src/react/shared/icon";
-import { IconType } from "src/shared/types";
+import { IconType } from "src/react/shared/icon/types";
 
 interface SortBubbleProps {
 	sortDir: SortDir;
