+++ src/shared/stringUtils.js
