+++ src/shared/stringUtils.ts
