+++ src/shared/types.ts
@@ -10,39 +10,3 @@ export enum Color {
 	PINK = "pink",
 	RED = "red",
 }
-
-export const enum IconType {
-	ARROW_UPWARD,
-	ARROW_DOWNWARD,
-	KEYBOARD_ARROW_UP,
-	KEYBOARD_ARROW_DOWN,
-	KEYBOARD_DOUBLE_ARROW_UP,
-	KEYBOARD_DOUBLE_ARROW_DOWN,
-	KEYBOARD_ARROW_LEFT,
-	KEYBOARD_ARROW_RIGHT,
-	KEYBOARD_DOUBLE_ARROW_LEFT,
-	KEYBOARD_DOUBLE_ARROW_RIGHT,
-	KEYBOARD_BACKSPACE,
-	DELETE,
-	MORE_VERT,
-	SORT,
-	MOVE_UP,
-	MOVE_DOWN,
-	TEXT_SNIPPET,
-	EDIT,
-	MORE_HORIZ,
-	CLOSE,
-	NOTES,
-	CHECK,
-	LABEL,
-	LIST,
-	NUMBERS,
-	CALENDAR_TODAY,
-	SCHEDULE,
-	SEARCH,
-	PAYMENTS,
-	TUNE,
-	ADD,
-	DRAG_INDICATOR,
-	VIEW_COLUMN,
-}
